requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
# psycopg2-binary==2.9.9  # Disabled due to compilation issues on macOS with Python 3.14
# dj-database-url==2.1.0  # Disabled due to psycopg2 dependency
# sqlite3  # Built-in Python module, no need to specify version

# CORS
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.0

# File handling
django-cleanup==8.0.0

# API integrations
requests==2.31.0
# google-generativeai==0.3.2  # Disabled due to Python 3.14 compatibility issues
# protobuf==4.21.12  # Fixed version for Python 3.14 compatibility

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Utilities
celery==5.3.6
redis==5.0.1
# django-celery-beat==2.6.0  # Disabled due to Django 5.0.1 compatibility issues

# Telegram Bot
python-telegram-bot==20.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0